requirements.txt
flask
Flask-RESTful==0.3.5
requests
spark-python